Implémentation Update / Delete / Create message => OK

+++ WalkForum.Application/Messages/Commands/DeleteMessage/DeleteMessageForPostCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿
 using MediatR;
 using Microsoft.Extensions.Logging;
+using WalkForum.Application.Users;
+using WalkForum.Domain.Constants;
 using WalkForum.Domain.Exceptions;
 using WalkForum.Domain.Repositories;
 
@@ -8,12 +10,14 @@ namespace WalkForum.Application.Messages.Commands.DeleteMessage;
 
 public class DeleteMessageForPostCommandHandler(ILogger<DeleteMessageForPostCommandHandler> logger, 
     IPostsRepository postsRepository,
-    IMessagesRepository messagesRepository
+    IMessagesRepository messagesRepository,
+    IUserContext userContext
     ) : IRequestHandler<DeleteMessageForPostCommand>
 {
     public async Task Handle(DeleteMessageForPostCommand request, CancellationToken cancellationToken)
     {
        logger.LogWarning("Removing message with id {messageId}, from post with id {postId}", request.MessageId, request.PostId);
+
 
         var post = await postsRepository.GetById(request.PostId);
         if (post is null) throw new NotFoundException("Post not found");
@@ -23,6 +27,7 @@ public class DeleteMessageForPostCommandHandler(ILogger<DeleteMessageForPostComm
 
         if (message.PostId != post.Id) throw new BadRequestException("Message not in this post");
 
+        if (message.UserId != userContext.GetCurrentUser().Id && !userContext.GetCurrentUser().IsInRole(UserRoles.Administrator) && !userContext.GetCurrentUser().IsInRole(UserRoles.Moderator)) throw new ForbiddenException("Not authorized");
         
         await messagesRepository.Delete(message);
